+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/internal/audio/DefaultAudioClientObserver.kt
@@ -83,7 +83,19 @@ class DefaultAudioClientObserver(
 
     override fun onAudioClientStateChange(newInternalAudioState: Int, newInternalAudioStatus: Int) {
         val newAudioStatus: MeetingSessionStatusCode? = toAudioStatus(newInternalAudioStatus)
-        var newAudioState: SessionStateControllerAction = toAudioClientState(newInternalAudioState, newAudioStatus)
+        val newAudioState: SessionStateControllerAction = toAudioClientState(newInternalAudioState, newAudioStatus)
+
+        if (newAudioStatus == MeetingSessionStatusCode.AudioInputDeviceNotResponding) {
+            val attributes = mutableMapOf<EventAttributeName, Any>(
+                EventAttributeName.deviceAccessErrorMessage to "Audio input device not responding"
+            )
+            eventAnalyticsController.publishEvent(EventName.deviceAccessFailed, attributes)
+        } else if (newAudioStatus == MeetingSessionStatusCode.AudioOutputDeviceNotResponding) {
+            val attributes = mutableMapOf<EventAttributeName, Any>(
+                EventAttributeName.deviceAccessErrorMessage to "Audio output device not responding"
+            )
+            eventAnalyticsController.publishEvent(EventName.deviceAccessFailed, attributes)
+        }
 
         if (newAudioStatus == null) {
             logger.warn(