+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/DefaultAudioVideoFacadeTest.kt
@@ -10,6 +10,8 @@ import android.content.Context
 import androidx.core.content.ContextCompat
 import com.amazonaws.services.chime.sdk.meetings.analytics.EventAnalyticsController
 import com.amazonaws.services.chime.sdk.meetings.analytics.EventAnalyticsObserver
+import com.amazonaws.services.chime.sdk.meetings.analytics.EventAttributeName
+import com.amazonaws.services.chime.sdk.meetings.analytics.EventName
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioDeviceCapabilities
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.activespeakerdetector.ActiveSpeakerDetectorFacade
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.contentshare.ContentShareController
@@ -34,8 +36,10 @@ import io.mockk.impl.annotations.MockK
 import io.mockk.mockkClass
 import io.mockk.mockkStatic
 import io.mockk.verify
+import kotlin.Any
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
+import org.junit.Assert.assertThrows
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
@@ -110,11 +114,17 @@ class DefaultAudioVideoFacadeTest {
     @Before
     fun setup() = MockKAnnotations.init(this, relaxUnitFun = true)
 
-    @Test(expected = SecurityException::class)
+    @Test
     fun `start should throw exception when the required permissions are not granted with default AudioDeviceCapabilities`() {
         mockkStatic(ContextCompat::class)
         every { ContextCompat.checkSelfPermission(any(), any()) } returns 1
-        audioVideoFacade.start()
+        assertThrows(SecurityException::class.java) {
+            audioVideoFacade.start()
+        }
+        val attributes = mutableMapOf<EventAttributeName, Any>(
+            EventAttributeName.deviceAccessErrorMessage to "No audio permission"
+        )
+        verify(exactly = 1) { eventAnalyticsController.publishEvent(EventName.deviceAccessFailed, attributes) }
     }
 
     @Test