+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## PENDING
+
+### Added
+* Added meetingReconnected meeting event
+* Added meetingReconnected/audioInputFailed/deviceAccessFailed meeting event
+* Added meetingReconnectDurationMs/audioInputErrorMessage/deviceAccessErrorMessage meeting event
+
 ## [0.25.0] - 2025-06-17
 
 ### Changed